# backend/docs/swagger.yaml
basePath: /
definitions:
  dto.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  dto.SignInRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.SignUpRequest:
    properties:
      address:
        type: string
      bank_account_name:
        type: string
      bank_account_number:
        type: string
      bank_name:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
    required:
    - address
    - bank_account_name
    - bank_account_number
    - bank_name
    - email
    - name
    - password
    - phone
    type: object
...
  dto.UpdateUserBankingRequest:
    properties:
      bank_account_name:
        type: string
      bank_account_number:
        description: Validate bank account number format (numeric and > 0)
        type: string
      bank_name:
        type: string
    type: object
  dto.UpdateUserProfileRequest:
    properties:
      address:
        type: string
      email:
        description: Validate email format
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  utils.GenericResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for your invoice app.
  title: Smart Invoice Generator API
  version: "1.0"
paths:
  /v1/protected/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh access token using a valid refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Refresh Token
      tags:
      - auth
...
  /v1/protected/me:
    get:
      description: Get details of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - auth
  /v1/protected/me/banking:
    put:
      consumes:
      - application/json
      description: Update user banking details
      parameters:
      - description: Update User Banking Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserBankingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Update User Banking
      tags:
      - auth
  /v1/protected/me/change-password:
    post:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Change Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - auth
  /v1/protected/me/deactivate:
    post:
      consumes:
      - application/json
      description: Deactivate user account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Deactivate User
      tags:
      - auth
  /v1/protected/me/profile:
    put:
      consumes:
      - application/json
      description: Update user profile details
      parameters:
      - description: Update User Profile Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - auth
  /v1/public/auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Sign In Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "403":
          description: Account deactivated - can be restored by registering again
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      summary: User Sign In
      tags:
      - auth
  /v1/public/auth/sign-up:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Sign Up Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.GenericResponse'
      summary: User Sign Up
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"